requirements/base.txt
wheel

pytz==2019.1  # https://github.com/stub42/pytz
python-slugify==3.0.3  # https://github.com/un33k/python-slugify
Pillow==6.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==19.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==4.1.3  # https://github.com/evansd/whitenoise
redis==3.3.11  # https://github.com/antirez/redis





#celery==4.3.0  # pyup: < 5.0  # https://github.com/celery/celery
celery==4.4.0rc4 # pyup: < 5.0  # https://github.com/celery/celery
# gevent
# eventlet


django-celery-beat==1.5.0  # https://github.com/celery/django-celery-beat
flower==0.9.3  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
# django==2.2.7  # pyup: < 3.0  # https://www.djangoproject.com/
django==2.2.19  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.39.1  # https://github.com/pennersr/django-allauth

django-redis==4.10.0  # https://github.com/niwinz/django-redis

# Django REST Framework
djangorestframework==3.10.3  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
# django-model-utils


# MMM
# abuse
requests
easyargs
click
future
git+https://github.com/LetMeR00t/abuse_finder.git@python3.6

google-api-python-client
google-auth
oauth2client<4.0.0

tldextract
dnspython
pythonwhois-alt

# another
django-rest-auth

django-pg-returning==1.2.2
celery_once


asyncio
aiohttp
aiodns

aiofiles

django-mathfilters
django-rest-swagger
djangorestframework-datatables
# django_tables2
# psutil
django-redis-sessions
django-sql-explorer
django-cacheops

# python-dotenv
django-phonenumber-field
phonenumbers==8.10.23
django-recaptcha


# django-mail-panel
beautifulsoup4


# django-screamshot
# webscreenshot
#git+git://github.com/maaaaz/webscreenshot.git@a491d50814b147088c3b293da7152b21620be5b8
future
argparse



psycopg2
# anymail
django-tinycontent
# -e git://github.com/cypreess/django-plans.git#egg=django-plans

# django-payments
django-related-admin
git+https://github.com/django-getpaid/django-plans.git
django-plans-payments

django-templated-email





# BUG!!!! https://github.com/django-crispy-forms/django-crispy-forms/pull/921
django-crispy-forms==1.8.1  # https://github.com/django-crispy-forms/django-crispy-forms
django-widget-tweaks
# sentry-sdk==0.13.4
# supervisor-win ; sys_platform == 'win32'

aiocontextvars
# django-background-tasks==1.2.4
# git+https://github.com/arteria/django-background-tasks.git
#django-suit==0.2.28
https://github.com/darklow/django-suit/tarball/v2
# https://github.com/darklow/django-suit/tarball/v2
validators

kombu==4.6.6


django-debug-toolbar==2.0
django-extensions==2.2.1

django-countries==6.1.2

python-interface
# django-dbtemplates
django-reversion
